pyproject.toml
[project]
name = "epl-mastermind"
version = "0.1.0"
description = "Premier League analytics lakehouse — FPL + football-data.org on a portable stack (dlt → DucLake → SQLMesh)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dlt[ducklake]>=1.20.0",
    "polars>=1.36.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlmesh[duckdb,github,web]>=0.228.1",
]
